dedicated time entry create service
